feat: create aggregator if not present; reformat code

+++ server/http/server.go
@@ -18,7 +18,7 @@ import (
 	nervexutil "go-sensephoenix.sensetime.com/nervex-operator/utils"
 )
 
-type NervexServer struct {
+type NerveXServer struct {
 	KubeClient         *kubernetes.Clientset
 	DynamicClient      dynamic.Interface
 	Log                logr.Logger
@@ -27,7 +27,7 @@ type NervexServer struct {
 	alconfig           string
 }
 
-type NervexJobRequest struct {
+type NerveXJobRequest struct {
 	Namespace   string           `json:"namespace"`
 	Coordinator string           `json:"coordinator"`
 	Actors      ResourceQuantity `json:"actors"`
@@ -41,7 +41,7 @@ type ResourceQuantity struct {
 	Memory   resource.Quantity `json:"memory"`
 }
 
-type NervexJobResponse struct {
+type NerveXJobResponse struct {
 	Namespace   string   `json:"namespace"`
 	Coordinator string   `json:"coordinator"`
 	Aggregator  string   `json:"aggregator"`
@@ -49,15 +49,15 @@ type NervexJobResponse struct {
 	Learners    []string `json:"learners"`
 }
 
-func NewNervexServer(
+func NewNerveXServer(
 	kubeClient *kubernetes.Clientset,
 	dynamicClient dynamic.Interface,
 	log logr.Logger,
 	alconfigDyInformer cache.SharedIndexInformer,
 	njDyInformer cache.SharedIndexInformer,
-	alconfig string) *NervexServer {
+	alconfig string) *NerveXServer {
 
-	return &NervexServer{
+	return &NerveXServer{
 		KubeClient:         kubeClient,
 		DynamicClient:      dynamicClient,
 		Log:                log,
@@ -67,8 +67,8 @@ func NewNervexServer(
 	}
 }
 
-func (s *NervexServer) Start() error {
-	log := s.Log.WithName("NervexServer")
+func (s *NerveXServer) Start() error {
+	log := s.Log.WithName("NerveXServer")
 	http.HandleFunc("/api/v1alpha1/add", s.Add)
 	http.HandleFunc("/api/v1alpha1/delete", s.Delete)
 	http.HandleFunc("/healthz", healthz)
@@ -80,11 +80,11 @@ func (s *NervexServer) Start() error {
 	return nil
 }
 
-func (s *NervexServer) Add(w http.ResponseWriter, r *http.Request) {
-	log := s.Log.WithName("NervexServer")
+func (s *NerveXServer) Add(w http.ResponseWriter, r *http.Request) {
+	log := s.Log.WithName("NerveXServer")
 
 	// parse request body
-	var njreq NervexJobRequest
+	var njreq NerveXJobRequest
 	err := json.NewDecoder(r.Body).Decode(&njreq)
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusBadRequest)
@@ -93,81 +93,90 @@ func (s *NervexServer) Add(w http.ResponseWriter, r *http.Request) {
 	log.Info("request body: ", "request", njreq)
 
 	// get ALConfig
-	obj, exists, err := s.alconfigDyInformer.GetIndexer().GetByKey(s.alconfig)
+	alconfig, err := s.getALConfig()
 	if err != nil {
-		errMsg := fmt.Sprintf("failed to get ALConfig: %s", err)
-		http.Error(w, errMsg, http.StatusInternalServerError)
-		return
-	}
-
-	if !exists {
-		http.Error(w, fmt.Sprintf("ActorLearnerConfig: %s not exists", s.alconfig), http.StatusInternalServerError)
-		return
-	}
-
-	alconfig := &nervexv1alpha1.ActorLearnerConfig{}
-	err = nervexutil.GetObjectFromUnstructured(obj, alconfig)
-	if err != nil {
+		log.Error(err, "failed to get ALConfig", "alconfig", s.alconfig)
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
 
+	// get ownReference of request coordinator
+	ownRefer, err := s.getOwnerReference(njreq)
+	if err != nil {
+		log.Error(err, "failed to get OwnerReference of coordinator %s", nervexutil.NamespacedName(njreq.Namespace, njreq.Coordinator))
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+
+	_, _, _, ag, err := s.listActorsAndLearners(njreq.Namespace, ownRefer.Name)
+	if err != nil {
+		log.Error(err, "failed to list actors and learners")
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
+	}
+	needAgg := false
+	if ag == nil {
+		needAgg = true
+	}
+
+	// create actors and learners
+	actors, learners, aggregator, err := s.createActorsAndLearnersFromALConfig(alconfig, &njreq, *ownRefer, needAgg)
+	if err != nil {
+		errMsg := fmt.Sprintf("failed create actors and learners: %s", err)
+		log.Error(err, "failed create actors and learners")
+		http.Error(w, errMsg, http.StatusInternalServerError)
+		return
+	}
+
+	// write response
+	writeResponse(w, njreq.Namespace, njreq.Coordinator, actors, learners, aggregator)
+}
+
+func (s *NerveXServer) getOwnerReference(njreq NerveXJobRequest) (*metav1.OwnerReference, error) {
 	// get coordinator
 	coordinator, err := s.KubeClient.CoreV1().Pods(njreq.Namespace).Get(context.Background(), njreq.Coordinator, metav1.GetOptions{})
 	if err != nil {
 		errMsg := fmt.Sprintf("failed to get coordinator: %s", err)
-		http.Error(w, errMsg, http.StatusInternalServerError)
-		return
+		return nil, fmt.Errorf(errMsg)
 	}
 
-	ownRefers := coordinator.GetOwnerReferences()
-	if len(ownRefers) == 0 {
-		errMsg := fmt.Sprintf("orphaned coordinator found: %s/%s, please ensure which NervexJob it belongs to", coordinator.Namespace, coordinator.Name)
-		http.Error(w, errMsg, http.StatusInternalServerError)
-		return
-	}
 	var ownRefer metav1.OwnerReference
+	ownRefers := coordinator.GetOwnerReferences()
+	ownByNerveX := false
 	for _, ref := range ownRefers {
-		if ref.Kind == "NerveXJob" {
+		if ref.Kind == nervexv1alpha1.KindNerveXJob {
 			ownRefer = ref
+			ownByNerveX = true
 		}
 	}
+	if !ownByNerveX {
+		errMsg := fmt.Sprintf("coordinator is not owned by NerveXJob")
+		return nil, fmt.Errorf(errMsg)
+	}
 
-	// get NervexJob
+	// get NerveXJob
 	njKey := fmt.Sprintf("%s/%s", njreq.Namespace, ownRefer.Name)
-	obj, exists, err = s.njDyInformer.GetIndexer().GetByKey(njKey)
+	_, exists, err := s.njDyInformer.GetIndexer().GetByKey(njKey)
 	if err != nil {
-		errMsg := fmt.Sprintf("failed to get NervexJob: %s", err)
-		http.Error(w, errMsg, http.StatusInternalServerError)
-		return
+		errMsg := fmt.Sprintf("failed to get NerveXJob: %s", err)
+		return nil, fmt.Errorf(errMsg)
 	}
 
 	if !exists {
-		http.Error(w, fmt.Sprintf("NerveXJob: %s not exists in cache", njKey), http.StatusInternalServerError)
-		return
+		errMsg := fmt.Sprintf("NerveXJob: %s not exists in cache", njKey)
+		return nil, fmt.Errorf(errMsg)
 	}
 
-	nj := &nervexv1alpha1.NerveXJob{}
-	err = nervexutil.GetObjectFromUnstructured(obj, nj)
-	if err != nil {
-		errMsg := fmt.Sprintf("failed to convert NervexJob: %s", err)
-		http.Error(w, errMsg, http.StatusInternalServerError)
-		return
-	}
+	return &ownRefer, nil
+}
 
-	// create actors and learners
-	actors, learners, err := s.createActorsAndLearnersFromALConfig(alconfig, &njreq, ownRefer)
-	if err != nil {
-		errMsg := fmt.Sprintf("failed create actors and learners: %s", err)
-		http.Error(w, errMsg, http.StatusInternalServerError)
-		return
-	}
-
-	rep := NervexJobResponse{
-		Namespace:   njreq.Namespace,
-		Coordinator: njreq.Coordinator,
+func writeResponse(w http.ResponseWriter, ns, coordinator string, actors, learners []string, aggregator string) {
+	rep := NerveXJobResponse{
+		Namespace:   ns,
+		Coordinator: coordinator,
 		Actors:      actors,
 		Learners:    learners,
+		Aggregator:  aggregator,
 	}
 	w.Header().Set("Conten-Type", "application/json")
 	w.WriteHeader(http.StatusOK)
@@ -184,10 +193,47 @@ func (s *NervexServer) Add(w http.ResponseWriter, r *http.Request) {
 	}
 }
 
-func (s *NervexServer) createActorsAndLearnersFromALConfig(
+func (s *NerveXServer) getALConfig() (*nervexv1alpha1.ActorLearnerConfig, error) {
+	obj, exists, err := s.alconfigDyInformer.GetIndexer().GetByKey(s.alconfig)
+	if err != nil {
+		errMsg := fmt.Sprintf("failed to get ALConfig: %s", err)
+		return nil, fmt.Errorf(errMsg)
+	}
+	if !exists {
+		errMsg := fmt.Sprintf("ActorLearnerConfig: %s not exists", s.alconfig)
+		return nil, fmt.Errorf(errMsg)
+	}
+
+	alconfig := &nervexv1alpha1.ActorLearnerConfig{}
+	err = nervexutil.GetObjectFromUnstructured(obj, alconfig)
+	if err != nil {
+		return nil, err
+	}
+
+	return alconfig, nil
+}
+
+func (s *NerveXServer) listActorsAndLearners(ns, jobName string) (
+	actors []*corev1.Pod, learners []*corev1.Pod, coordinator *corev1.Pod, aggregator *corev1.Pod, err error) {
+	// list pods that belong to the NerveXJob
+	pods, err := s.listPods(ns, jobName)
+	if err != nil {
+		return
+	}
+
+	// classify pods
+	actors, learners, coordinator, aggregator, err = nervexutil.ClassifyPods(pods)
+	if err != nil {
+		return
+	}
+	return
+}
+
+func (s *NerveXServer) createActorsAndLearnersFromALConfig(
 	alconfig *nervexv1alpha1.ActorLearnerConfig,
-	njreq *NervexJobRequest,
-	ownRefer metav1.OwnerReference) ([]string, []string, error) {
+	njreq *NerveXJobRequest,
+	ownRefer metav1.OwnerReference,
+	needAgg bool) ([]string, []string, string, error) {
 
 	// create actors
 	actorTemplate := alconfig.Spec.Actor.Template
@@ -195,7 +241,7 @@ func (s *NervexServer) createActorsAndLearnersFromALConfig(
 		nervexutil.ActorName, nervexutil.DefaultActorContainerName, nervexutil.DefaultActorPortName, nervexutil.DefaultActorPort)
 
 	if err != nil {
-		return nil, nil, err
+		return nil, nil, "", err
 	}
 
 	// create learners
@@ -204,30 +250,46 @@ func (s *NervexServer) createActorsAndLearnersFromALConfig(
 		nervexutil.LearnerName, nervexutil.DefaultLearnerContainerName, nervexutil.DefaultLearnerPortName, nervexutil.DefaultLearnerPort)
 
 	if err != nil {
-		return nil, nil, err
+		return nil, nil, "", err
 	}
 
-	return actors, learners, nil
+	if !needAgg {
+		return actors, learners, "", nil
+	}
+	// create aggregator
+	aggregatorTemplate := alconfig.Spec.Learner.Template
+	aggregators, err := s.createPodsAndServices(&aggregatorTemplate, ownRefer, njreq,
+		nervexutil.AggregatorName, nervexutil.DefaultAggregatorContainerName, nervexutil.DefaultAggregatorPortName, nervexutil.DefaultAggregatorPort)
+	if err != nil {
+		return nil, nil, "", err
+	}
+	aggregator := aggregators[0]
+
+	return actors, learners, aggregator, nil
 }
 
-func (s *NervexServer) createPodsAndServices(
+func (s *NerveXServer) createPodsAndServices(
 	template *corev1.PodTemplateSpec,
 	ownRefer metav1.OwnerReference,
-	njreq *NervexJobRequest,
+	njreq *NerveXJobRequest,
 	replicaType, containerName, portName string, defaultPort int32) ([]string, error) {
 
-	log := s.Log.WithName("NervexServer")
+	log := s.Log.WithName("NerveXServer")
 	coordinator := njreq.Coordinator
 	ns := njreq.Namespace
 
 	portEnv := ""
 	resources := ResourceQuantity{}
-	if replicaType == nervexutil.ActorName {
+	switch replicaType {
+	case nervexutil.ActorName:
 		resources = njreq.Actors
 		portEnv = "ACTOR_PORT"
-	} else if replicaType == nervexutil.LearnerName {
+	case nervexutil.LearnerName:
 		resources = njreq.Learners
 		portEnv = "LEARNER_PORT"
+	case nervexutil.AggregatorName:
+		resources = ResourceQuantity{Replicas: 1}
+		portEnv = "AGGREGATOR_PORT"
 	}
 
 	// add labels to pod template
@@ -245,7 +307,7 @@ func (s *NervexServer) createPodsAndServices(
 	port, ok := nervexutil.GetPortFromPodTemplate(template, containerName, portName)
 	if !ok {
 		port = defaultPort
-		log.Info("no port found, use default port", "port", port)
+		log.Info("no port found, use default port", "container", containerName, "port", port)
 		nervexutil.SetPodTemplatePort(template, containerName, portName, port)
 	}
 
@@ -306,16 +368,18 @@ func SetPodTemplateResources(template *corev1.PodTemplateSpec, resources Resourc
 			template.Spec.Containers[i].Resources.Limits[corev1.ResourceMemory] = resources.Memory
 			template.Spec.Containers[i].Resources.Requests[corev1.ResourceMemory] = resources.Memory
 		}
-		template.Spec.Containers[i].Resources.Limits[corev1.ResourceName("nvidia.com/gpu")] = resources.Gpu
-		template.Spec.Containers[i].Resources.Requests[corev1.ResourceName("nvidia.com/gpu")] = resources.Gpu
+		if !resources.Gpu.IsZero() {
+			template.Spec.Containers[i].Resources.Limits[corev1.ResourceName("nvidia.com/gpu")] = resources.Gpu
+			template.Spec.Containers[i].Resources.Requests[corev1.ResourceName("nvidia.com/gpu")] = resources.Gpu
+		}
 	}
 }
 
-func (s *NervexServer) Delete(w http.ResponseWriter, r *http.Request) {
-	log := s.Log.WithName("NervexServer")
+func (s *NerveXServer) Delete(w http.ResponseWriter, r *http.Request) {
+	log := s.Log.WithName("NerveXServer")
 
 	// parse request body
-	var njreq NervexJobRequest
+	var njreq NerveXJobRequest
 	err := json.NewDecoder(r.Body).Decode(&njreq)
 	if err != nil {
 		http.Error(w, err.Error(), http.StatusBadRequest)
@@ -323,62 +387,18 @@ func (s *NervexServer) Delete(w http.ResponseWriter, r *http.Request) {
 	}
 	log.Info("delete request body: ", "request", njreq)
 
-	// get coordinator
-	coordinator, err := s.KubeClient.CoreV1().Pods(njreq.Namespace).Get(context.Background(), njreq.Coordinator, metav1.GetOptions{})
+	// get ownReference of the request coordinator
+	ownRefer, err := s.getOwnerReference(njreq)
 	if err != nil {
-		errMsg := fmt.Sprintf("failed to get coordinator: %s", err)
-		http.Error(w, errMsg, http.StatusInternalServerError)
+		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
 
-	ownRefers := coordinator.GetOwnerReferences()
-	if len(ownRefers) == 0 {
-		errMsg := fmt.Sprintf("orphaned coordinator found: %s/%s, please ensure which NervexJob it belongs to", coordinator.Namespace, coordinator.Name)
-		http.Error(w, errMsg, http.StatusInternalServerError)
-		return
-	}
-	var ownRefer metav1.OwnerReference
-	for _, ref := range ownRefers {
-		if ref.Kind == "NerveXJob" {
-			ownRefer = ref
-		}
-	}
-
-	// get NervexJob
-	njKey := fmt.Sprintf("%s/%s", njreq.Namespace, ownRefer.Name)
-	obj, exists, err := s.njDyInformer.GetIndexer().GetByKey(njKey)
+	// list pods that belong to the NerveXJob
+	actors, learners, _, _, err := s.listActorsAndLearners(njreq.Namespace, ownRefer.Name)
 	if err != nil {
-		errMsg := fmt.Sprintf("failed to get NervexJob: %s", err)
-		http.Error(w, errMsg, http.StatusInternalServerError)
-		return
-	}
-
-	if !exists {
-		http.Error(w, fmt.Sprintf("NerveXJob: %s not exists in cache", njKey), http.StatusInternalServerError)
-		return
-	}
-
-	nj := &nervexv1alpha1.NerveXJob{}
-	err = nervexutil.GetObjectFromUnstructured(obj, nj)
-	if err != nil {
-		errMsg := fmt.Sprintf("failed to convert NervexJob: %s", err)
-		http.Error(w, errMsg, http.StatusInternalServerError)
-		return
-	}
-
-	// list pods
-	pods, err := s.listPods(njreq.Namespace, ownRefer)
-	if err != nil {
-		errMsg := fmt.Sprintf("failed to list pod through ownrefernece: %s", err)
-		http.Error(w, errMsg, http.StatusInternalServerError)
-		return
-	}
-
-	// classify pods
-	actors, learners, _, err := nervexutil.ClassifyPods(pods)
-	if err != nil {
-		errMsg := fmt.Sprintf("unable to classify pods: %s", err)
-		http.Error(w, errMsg, http.StatusInternalServerError)
+		log.Error(err, "failed to list actors and learners")
+		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
 	}
 
@@ -398,30 +418,13 @@ func (s *NervexServer) Delete(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	rep := NervexJobResponse{
-		Namespace:   njreq.Namespace,
-		Coordinator: njreq.Coordinator,
-		Actors:      delActors,
-		Learners:    delLearners,
-	}
-	w.Header().Set("Conten-Type", "application/json")
-	w.WriteHeader(http.StatusOK)
-	repJson, err := json.Marshal(rep)
-	if err != nil {
-		errMsg := fmt.Sprintf("failed to marshal json: %s", err)
-		http.Error(w, errMsg, http.StatusInternalServerError)
-		return
-	}
-	_, err = w.Write(repJson)
-	if err != nil {
-		errMsg := fmt.Sprintf("failed to write json: %s", err)
-		http.Error(w, errMsg, http.StatusInternalServerError)
-	}
+	// write response
+	writeResponse(w, njreq.Namespace, njreq.Coordinator, delActors, delLearners, "")
 }
 
-func (s *NervexServer) listPods(ns string, ownRefer metav1.OwnerReference) ([]*corev1.Pod, error) {
+func (s *NerveXServer) listPods(ns string, jobName string) ([]*corev1.Pod, error) {
 	labelSelector, err := metav1.LabelSelectorAsSelector(&metav1.LabelSelector{
-		MatchLabels: nervexutil.GenLabels(ownRefer.Name),
+		MatchLabels: nervexutil.GenLabels(jobName),
 	})
 	if err != nil {
 		return nil, err
@@ -440,18 +443,23 @@ func (s *NervexServer) listPods(ns string, ownRefer metav1.OwnerReference) ([]*c
 	return pods, nil
 }
 
-func (s *NervexServer) DeletePodsAndServices(pods []*corev1.Pod, njreq *NervexJobRequest, replicaType string) ([]string, error) {
+func (s *NerveXServer) DeletePodsAndServices(pods []*corev1.Pod, njreq *NerveXJobRequest, replicaType string) ([]string, error) {
 	results := []string{}
 	resources := ResourceQuantity{}
 	ns := njreq.Namespace
 
-	if replicaType == nervexutil.ActorName {
+	switch replicaType {
+	case nervexutil.ActorName:
 		resources = njreq.Actors
-	} else if replicaType == nervexutil.LearnerName {
+	case nervexutil.LearnerName:
 		resources = njreq.Learners
 	}
 
 	for _, pod := range pods {
+		// break if enough
+		if len(results) >= resources.Replicas {
+			break
+		}
 
 		err := s.KubeClient.CoreV1().Pods(njreq.Namespace).Delete(context.Background(), pod.Name, metav1.DeleteOptions{})
 		if err != nil {
@@ -464,10 +472,7 @@ func (s *NervexServer) DeletePodsAndServices(pods []*corev1.Pod, njreq *NervexJo
 
 		result := fmt.Sprintf("%s.%s", pod.Name, ns)
 		results = append(results, result)
-		// break if enough
-		if len(results) == resources.Replicas {
-			break
-		}
+
 	}
 
 	return results, nil
